Fix the Options Update: - During Update , the Options must be aligned with the way it has been created - use dict() in order to copy and not reference Devices for setting DOptions in z_command, z_domoticz and z_heartbeat
Python never implicitly copies objects. When you set dict2 = dict1, you are making them refer to the same exact dict object, so when you mutate it, all references to it keep referring to the object in its current state.

If you want to copy the dict (which is rare), you have to do so explicitly with